Reporte ordenes por facturar, totales por moneda
